feat: extract params of pdf
